RTC: LPC32xx: Improve RTC invalid state detection
When a power cut of few seconds occurs (time depending on the board power
stage), RTC SRAM can be corrupted. When the RTC voltage starts to deteriorate,
bits in the RTC register change state. But if the key register stays valid and
other bits in the RTC like the up/down time change and the system is repowered
on with the key register still valid, the software will think the RTC is ok.

In these cases, you can see some weird dates like 2035, 1982, ...

This patch uses the down counter value to validate the up counter one (it must
be equal to u32 inverted up counter value). If values don't match, it reset
both counters.

Signed-off by: Martin Chaplet <m.chaplet@kernlink.fr>