lpc32x0: nand: Workaround for NAND data loss issue
Temporary workaround for DMA/NAND support. There seems to be an issue when
using dma_map_single for NAND buffers in very rare cases. When this issue
happens, the NAND data will be corrupted. This tends to happen only when
under heavy NAND load. For now, keep this define enabled.